==> 05_ChlA_difference_0m_vs_5m.Rmd <==
---
title: "Økokyst data - ChlA measured at 0 m vs 5 m"
output: 
  html_document: 
    keep_md: yes
    toc: yes
---

## Library
```{r}
library(tidyverse)
library(lubridate)
```

## Data
Saved in script 02
```{r}
df_chla <- readRDS("Data/02_df_chla.RData")
df_stations <- readRDS("Data/02_df_stations.RData")
```

## Make data set for plots
Pick 0 and 5 m and rearrang
```{r}
df_chla2 <- df_chla %>% 
  filter(Depth >= 0 & Depth <= 5) %>%
  spread(Depth, KlfA) %>%
  rename(Depth_00_m = `0`, Depth_05_m = `5`) %>%
  mutate(Abs_difference_00m_vs_05m = Depth_00_m - Depth_05_m,
         Perc_difference_00m_vs_05m = 100*(Depth_00_m - Depth_05_m)/Depth_05_m,
         StationName = factor(StationName, levels = df_stations$StationName))
# Adding 'StationCodeName' for graphs
df_chla$StationCodeName  <- with(df_chla, paste(StationCode, StationName))
df_chla2$StationCodeName  <- with(df_chla2, paste(StationCode, StationName))
# Checking (line 1 should be 2x line 2)
df_chla %>% filter(Depth <= 5) %>% nrow()
df_chla2 %>% nrow()
```

### Summary of data
```{r, echo=FALSE, warning=FALSE}
cat("Summary of data\n")
df_chla2 %>% 
  select(Depth_00_m, Depth_05_m, Abs_difference_00m_vs_05m, Perc_difference_00m_vs_05m) %>%
  summary()
cat("Abs_difference_00m_vs_05m, percentiles\n")
df_chla2$Abs_difference_00m_vs_05m %>% quantile(seq(0,1,0.2), na.rm = TRUE)
cat("Abs_difference_00m_vs_05m, percentiles of absolute values\n")
df_chla2$Abs_difference_00m_vs_05m %>% abs() %>% quantile(seq(0,1,0.2), na.rm = TRUE)
cat("Perc_difference_00m_vs_05m, percentiles\n")
df_chla2$Perc_difference_00m_vs_05m %>% quantile(seq(0,1,0.2), na.rm = TRUE)
cat("Perc_difference_00m_vs_05m, percentiles of absolute values\n")
df_chla2$Perc_difference_00m_vs_05m %>% abs() %>% quantile(seq(0,1,0.2), na.rm = TRUE)
```

## Absolute difference
### Absolute difference, 5 m vs 0 m {.tabset}

#### Full x and y scales
```{r, fig.height=7, fig.width=9, echo=FALSE, warning=FALSE}
absdiff_max <- df_chla2$Abs_difference_00m_vs_05m %>% abs() %>% max() %>% ceiling()
absdiff_breaks <- c(0,0.1,0.25,0.5,1,2, absdiff_max)
absdiff_colors <- c("green4","green","orange","orange3","red","red4")
gg <- ggplot(df_chla2, 
             aes(x = Depth_00_m, 
                 y = Depth_05_m, 
                 color = cut(abs(Abs_difference_00m_vs_05m), 
                             absdiff_breaks, include.lowest = TRUE)
                 )
             ) +
  geom_point() +
  scale_color_manual("Absolute\ndifference", values = absdiff_colors, na.value = "grey30") +
  geom_abline(aes(intercept = 0, slope = 1)) +
  facet_wrap(~StationCodeName)
gg
```

#### Restricted x and y scales (not showing all points)
```{r, fig.height=7, fig.width=9, echo=FALSE, warning=FALSE}
gg + 
  coord_cartesian(xlim = c(0, 3.5), ylim = c(0, 4))
```

### Absolute difference, time series {.tabset}
#### Full y scale
```{r, fig.width=9, echo=FALSE, warning=FALSE}
gg <- ggplot(df_chla2, aes(Time, Abs_difference_00m_vs_05m)) +
  geom_hline(aes(yintercept = 0)) +
  geom_line(color = "brown", size = 1) +
  facet_wrap(~StationCodeName)
gg
```

#### Restricted y scale
```{r, fig.width=9, echo=FALSE, warning=FALSE}
gg + coord_cartesian(ylim = c(-2.5, 2.5))
```

### Time series of Chl a at 5 m, absolute difference as colour {.tabset}
  * Co loured/connected dots: Chl a at 5 m  
  * Blue crosses: Chl a at 0 m
  
#### Full y scale
```{r, fig.height=7, fig.width=9, echo=FALSE, warning=FALSE}
gg <- ggplot(df_chla2, 
             aes(x = Time, 
                 y = Depth_05_m, 
                 color = cut(abs(Abs_difference_00m_vs_05m), 
                             absdiff_breaks, include.lowest = TRUE))
       ) +
  geom_line(color = "black") +
  geom_point() +
  geom_point(data = df_chla2, aes(Time, Depth_00_m), color = "blue", shape = 4) +
  scale_color_manual("Absolute\ndifference", values = absdiff_colors, na.value = "grey30") +
  facet_wrap(~StationCodeName)
gg
```

#### Restricted y scale
```{r, fig.width=9, echo=FALSE, warning=FALSE}
gg + coord_cartesian(ylim = c(0, 4.5))
```

## Percent difference
### Percent difference, 5 m vs 0 m {.tabset}

#### Full x and y scales
```{r, fig.height=7, fig.width=9, echo=FALSE, warning=FALSE}
percent_max <- df_chla2$Perc_difference_00m_vs_05m %>% abs() %>% max() %>% ceiling()
percent_breaks <- c(0,10,20,50,100,300, percent_max)
percent_colors <- c("green3","green","orange","orange3","red","red4")
gg <- ggplot(df_chla2, aes(Depth_00_m, Depth_05_m, 
                     color = cut(abs(Perc_difference_00m_vs_05m), percent_breaks, include.lowest = TRUE))
       ) +
  geom_point() +
  scale_color_manual("Percent\ndifference", values = percent_colors, na.value = "grey30") +
  geom_abline(aes(intercept = 0, slope = 1)) +
  facet_wrap(~StationCodeName)
gg
```


#### Restricted x and y scales (not showing all points)
```{r, fig.height=7, fig.width=9, echo=FALSE, warning=FALSE}
gg + 
  coord_cartesian(xlim = c(0, 3.5), ylim = c(0, 4))
```

### Percent difference, time series {.tabset}

#### Full y scale
```{r, fig.width=9, echo=FALSE, warning=FALSE}
ggplot(df_chla2, aes(Time, Perc_difference_00m_vs_05m)) +
  geom_line() +
  facet_wrap(~StationCodeName)

```

### Time series of Chl a at 5 m, percent difference as colour {.tabset}
  * Coloured/connected dots: Chl a at 5 m  
  * Blue crosses: Chl a at 0 m

#### Full y scale
```{r, fig.height=7, fig.width=9, echo=FALSE, warning=FALSE}
gg <- ggplot(df_chla2, aes(Time, Depth_05_m, 
                       color = cut(abs(Perc_difference_00m_vs_05m), percent_breaks, include.lowest = TRUE))
       ) +
  geom_line(color = "black") +
  geom_point() +
  geom_point(data = df_chla2, aes(Time, Depth_00_m), color = "blue", shape = 4) +
  scale_color_manual("Percent\ndifference", values = percent_colors, na.value = "grey30") +
  facet_wrap(~StationCodeName)
gg
```

#### Restricted y scale
```{r, fig.height=7, fig.width=9, echo=FALSE, warning=FALSE}
gg + 
  coord_cartesian(ylim = c(0, 4.5))
```
  
# Absolute difference, plots for each station   
### Define stations  
```{r}
# unique(df_chla2$StationCodeName) %>% dput()
stations_oceanic <- c("VT70 Bjørnafjorden", 
                   "VT71 Skinnbrokleia",
                   "VR31 Tilremsfjorden", 
                   "VT3 Torbjørnskjær")

stations_fjord <- c("VT75 Fusafjorden", 
  "VT52 Kvinnheradsfjorden", "VT16 Kyrkjebø", "VT74 Maurangerfjorden", 
  "VT79 Nærsnes", "VT53 Tveitneset", "VR51 Korsen",
  "VR52 Broemsneset", "VT42 Korsfjorden_Trønd",  
  "VT2 Bastø", "VT10 Breiangen", "VT66 Håøyfjorden", 
  "VT67 Langesundsfjorden",   "VT65 Missingene")
```
  
### Scatter plot (0 m vs 5 m), function  
```{r, fig.height=5, fig.width=5, echo=FALSE, warning=FALSE}
plot_scatter <- function(codename){
  df_chla2_sel <- df_chla2 %>% 
    filter(StationCodeName == codename & !is.na(Abs_difference_00m_vs_05m))
  absdiff_max <- df_chla2_sel$Abs_difference_00m_vs_05m %>% 
    abs() %>% max() %>% ceiling()
  if (absdiff_max > 2){
    absdiff_breaks <- c(0,0.1,0.25,0.5,1,2, absdiff_max)
    absdiff_colors <- c("green4","green","orange","orange3","red","red4")
  } else {
    absdiff_breaks <- c(0,0.1,0.25,0.5,1,2)
    absdiff_colors <- c("green4","green","orange","orange3","red")
  }
  gg <- ggplot(
    df_chla2_sel, 
    aes(x = Depth_00_m, 
        y = Depth_05_m, 
        color = cut(abs(Abs_difference_00m_vs_05m), 
                    absdiff_breaks, include.lowest = TRUE)
    )) +
    geom_point(size = 2) +
    
    scale_color_manual("Absolute\ndifference", values = absdiff_colors, na.value = "grey30") +
    geom_abline(aes(intercept = 0, slope = 1), size = 1) +
    #?geom_hline(aes(yintercept = c(3.5,7,11,20))) +
    geom_vline(xintercept = c(3.5,7), linetype = "dashed") +
    geom_hline(yintercept = c(3.5,7), linetype = "dashed") +
    scale_x_continuous(breaks = seq(0, 20, 1)) +
    scale_y_continuous(breaks = seq(0, 20, 1)) +
    expand_limits(x = 0, y = 0) +
    coord_fixed() +
    labs(title = codename, x = "Chl a, 0 m", y = "Chl a, 5 m")
  gg
}

save_scatter <- function(codename, subfolder){
  gg <- plot_scatter(codename)
  ggsave(sprintf("Figures_0m_vs_5m/%s/%s Scatter.png", subfolder, codename), gg, width = 5, height = 4, dpi = 500)
}

# Test:
# plot_scatter("VR31 Tilremsfjorden")
```

#### Scatter, oceanic stations
```{r, fig.height=5, fig.width=5}
# For showing in notebook
# stations_oceanic %>% walk(~print(plot_scatter(.)))

# For saving
stations_oceanic %>% walk(~save_scatter(., "Oceanic"))
```

#### Scatter, closed stations
```{r, fig.width=5, fig.height=5}
stations_fjord %>% walk(~save_scatter(., "Fjord"))
```

### Difference vs. time, function
```{r, fig.height=4, fig.width=5, echo=FALSE, warning=FALSE}
plot_diff <- function(codename){
  df_chla2_sel <- df_chla2 %>% 
    filter(StationCodeName == codename & !is.na(Abs_difference_00m_vs_05m))
  absdiff_max <- df_chla2_sel$Abs_difference_00m_vs_05m %>% 
    abs() %>% max() %>% ceiling()
  if (absdiff_max > 2){
    absdiff_breaks <- c(0,0.1,0.25,0.5,1,2, absdiff_max)
    absdiff_colors <- c("green4","green","orange","orange3","red","red4")
  } else {
    absdiff_breaks <- c(0,0.1,0.25,0.5,1,2)
    absdiff_colors <- c("green4","green","orange","orange3","red")
  }
  gg <- ggplot(df_chla2_sel, 
             aes(x = Time, 
                 y = Abs_difference_00m_vs_05m, 
                 color = cut(abs(Abs_difference_00m_vs_05m), 
                             absdiff_breaks, include.lowest = TRUE))
       ) +
  geom_line(color = "black") +
  geom_point(size = 3) +
  scale_color_manual("Absolute\ndifference", values = absdiff_colors, na.value = "grey30") +
  scale_y_continuous(breaks = seq(-10, 10, 1)) +
  geom_hline(yintercept = 0) +
  expand_limits(y = c(-1,1)) +
  labs(title = codename, x = "Time", y = "Chl a difference (0 m minus 5 m)")
  gg
  }
plot_diff("VR31 Tilremsfjorden")

save_diff <- function(codename, subfolder){
  gg <- plot_diff(codename)
  ggsave(sprintf("Figures_0m_vs_5m/%s/%s Diff.png", subfolder, codename), gg,     width = 5, height = 4, dpi = 500)
}
# save_diff("VR31 Tilremsfjorden", "Oceanic")
```
#### Difference vs. time, oceanic stations
```{r, fig.height=5, fig.width=5}
stations_oceanic %>% walk(~save_diff(., "Oceanic"))
```

#### Difference vs. time, fjord stations
```{r, fig.width=5, fig.height=5}
stations_fjord %>% walk(~save_diff(., "Fjord"))
```


### Time series plot, function
```{r, fig.height=4, fig.width=5, echo=FALSE, warning=FALSE}
plot_time <- function(codename){
  df_chla2_sel <- df_chla2 %>% 
    filter(StationCodeName == codename & !is.na(Abs_difference_00m_vs_05m))
  absdiff_max <- df_chla2_sel$Abs_difference_00m_vs_05m %>% 
    abs() %>% max() %>% ceiling()
  if (absdiff_max > 2){
    absdiff_breaks <- c(0,0.1,0.25,0.5,1,2, absdiff_max)
    absdiff_colors <- c("green4","green","orange","orange3","red","red4")
  } else {
    absdiff_breaks <- c(0,0.1,0.25,0.5,1,2)
    absdiff_colors <- c("green4","green","orange","orange3","red")
  }
  gg <- ggplot(df_chla2_sel, 
             aes(x = Time, 
                 y = Depth_05_m, 
                 color = cut(abs(Abs_difference_00m_vs_05m), 
                             absdiff_breaks, include.lowest = TRUE))
       ) +
  geom_line(color = "black") +
  geom_point(size = 3) +
  geom_point(data = df_chla2_sel, aes(Time, Depth_00_m), color = "blue", shape = 4, size = 3) +
  scale_color_manual("Absolute\ndifference", values = absdiff_colors, na.value = "grey30") +
  scale_y_continuous(breaks = seq(0, 20, 1)) +
  geom_hline(yintercept = c(3.5,7), linetype = "dashed") +
  expand_limits(y = 0) +
  labs(title = codename, x = "Time", y = "Chl a")
  gg
  }
# plot_time("VR31 Tilremsfjorden")

save_time <- function(codename, subfolder){
  gg <- plot_time(codename)
  ggsave(sprintf("Figures_0m_vs_5m/%s/%s Time.png", subfolder, codename), gg,     width = 5, height = 4, dpi = 500)
}
# save_time("VR31 Tilremsfjorden", "Oceanic")
```

#### Time series plots, oceanic stations
```{r, fig.height=5, fig.width=5}
stations_oceanic %>% walk(~save_time(., "Oceanic"))
```

#### Time series plots, fjord stations
```{r, fig.width=5, fig.height=5}
stations_fjord %>% walk(~save_time(., "Fjord"))
```

### "Stack" plot (0 m, 5 m and 10 m), function
Using only March-October data
```{r, fig.height=5, fig.width=5, echo=FALSE, warning=FALSE}
plot_stack_makedata <- function(codename){
  df_chla_sel <- df_chla %>% 
    filter(StationCodeName == codename & !is.na(KlfA)) %>%
    filter(Depth %in% c(0,5,10) & month(Time) >= 3 & month(Time) <= 10) %>%
    group_by(Depth) %>%
    mutate(Rank = dense_rank(-KlfA), Second_highest = (Rank == 2))
  df_chla_sel
  }
# plot_stack_makedata("VR31 Tilremsfjorden")

plot_stack <- function(codename){
  df <- plot_stack_makedata(codename)
  gg <- ggplot(df, 
    aes(x = Depth, y = KlfA, color = Second_highest, size = factor(Second_highest))) +
    geom_point(shape = "_") +
    geom_hline(yintercept = c(3.5,7), linetype = "dashed") +
    scale_color_manual(values = c("grey40", "red3")) +
    scale_size_manual(values = c(8,14)) +
    scale_x_continuous(breaks = c(0, 5, 10)) +
    scale_y_continuous(breaks = seq(0, 20, 1)) +
    coord_cartesian(xlim = c(-2, 12)) +
    expand_limits(y = 0) +
    labs(title = codename, x = "Depth (m)", y = "Chl a") +
    theme(legend.position = "none")
  gg
}

save_stack <- function(codename, subfolder){
  gg <- plot_stack(codename)
  fn <- sprintf("Figures_0m_5m_10m/%s/%s Stack.png", subfolder, codename)
  ggsave(fn, gg, width = 5, height = 4, dpi = 400)
}

# Test:
# plot_stack("VR31 Tilremsfjorden")
# gg <- plot_stack("VR31 Tilremsfjorden")
# debugonce(save_stack)
# save_stack("VR31 Tilremsfjorden", "Oceanic")
```

#### Stack plots, oceanic stations
```{r, fig.height=5, fig.width=5}
# For showing in notebook
# stations_oceanic %>% walk(~print(plot_scatter(.)))

# For saving
stations_oceanic %>% walk(~save_stack(., "Oceanic"))
```

#### Stack plots, fjord stations
```{r, fig.width=5, fig.height=5}
stations_fjord %>% walk(~save_stack(., "Fjord"))
```
