==> 14_QA_CTD_2020_NorskehavetSor2.Rmd <==
---
title: "QA for CTDs in Norskehavet Sør I - ØKOKYST 2019 report"
author: "DHJ"
date: "13 1 2020"
output:
  html_document:
    keep_md: true
    toc: true
    toc_float: true

---

## QC done January 2021    
- CTDs with salinity + temp + oxygen      
- See mail from Caroline Mengeot <caroline.mengeot@niva.no> tirsdag 28. januar 2020 10:33    
- Collectors / stations:  
    * OKOKYST_NH_Sor2_Aquakompetanse: VR52
    * OKOKYST_NH_Sor2_MagneAuren: VT42  
    * OKOKYST_NH_Sor2_SNO: VR31
```{r, echo=FALSE, results='hide', message=FALSE, warning=FALSE}

## 1. Libraries

library(dplyr)
library(purrr)
library(ggplot2)
library(lubridate)
library(readxl)
library(tidyr)
library(RColorBrewer)
library(knitr)
library(kableExtra)

# library(niRvana)

source("12_QA_2019_from_excel_functions.R")
source("13_QA_CTD_2019_functions.R")

# RColorBrewer::display.brewer.all()

knitr::opts_chunk$set(echo = FALSE, results = 'hold')

```


## 1. Read files   

### New data (2020)  
- Read from excel files in folders under `K:/Avdeling/214-Oseanografi/DATABASER/OKOKYST_2017s`  
- 3 files in 3 separate folders in this case    
```{r, echo = FALSE, results = 'hide'}

# datafolder <- "K:/Prosjekter/!!! Tiltaksrettet industriovervåkning iht. vannforskriften/16371 - Yara Glomfjorden (tidl.14380)/_2019/Vannsøyle/xlsbase/TilAquamonitor"

# dir(datafolder)

read_excel_files <- FALSE

if (read_excel_files){
  
  # This was copied to C:
  # datafolder1 <- "K:/Avdeling/214-Oseanografi/DATABASER/OKOKYST_2017/OKOKYST_NH_Sor2_Aquakompetanse/xlsbase"
  datafolder1 <- "Datasett/OKOKYST_NH_Sor2_Aquakompetanse/xlsbase"
  fn <- paste0(datafolder1, "/", "VR52_2020_CTD.xlsm")
  # Check top of file:
  # read_excel(fn, sheet = "Data", n_max = 200)
  dat1 <- read_excel_droplines(fn, sheet = "Data", first_data_row = 3, cut_columns = TRUE)
  dat_stations1 <- read_excel(fn, sheet = "Stations")
  
  # As above, copied from K:/Avdeling/214-Oseanografi/DATABASER to C:
  datafolder2 <- "Datasett/OKOKYST_NH_Sor2_MagneAuren/xlsbase"
  fn <- paste0(datafolder2, "/", "VT42_2020_CTD.xlsm")
  dat2 <- read_excel_droplines(fn, sheet = "Data", first_data_row = 3, cut_columns = TRUE)
  # We don't need column 10 (empty) and 11 (Method)
  # dat2 <- dat2[1:9]
  dat_stations2 <- read_excel(fn, sheet = "Stations")

  # As above, copied from K:/Avdeling/214-Oseanografi/DATABASER to C:
  datafolder3 <- "Datasett/OKOKYST_NH_Sor2_SNO/xlsbase"
  fn <- paste0(datafolder3, "/", "VR31_CTD_2020.xlsm")
  dat3 <- read_excel_droplines(fn, sheet = "Data", first_data_row = 3, cut_columns = TRUE)
  dat_stations3 <- read_excel(fn, sheet = "Stations")

    
  # Combine data
  dat <- bind_rows(dat1, dat2, dat3)
  
  # Fix the first column names
  # NOT NEEDED
  
  # Add station names
  dat_stations <- bind_rows(dat_stations1, dat_stations2, dat_stations3) %>%
    rename(StationCode = StationName)

  dat <- dat %>%
    left_join(dat_stations %>% select(-ProjectId), by = "StationId")
  
  saveRDS(dat, "Data/14_QA_CTD_2020_NorskehavSor2_dat.rds")
  
} else {
  
  dat <- readRDS("Data/14_QA_CTD_2020_NorskehavSor2_dat.rds")
  
}

xtabs(~addNA(StationCode), dat)
xtabs(~addNA(Date), dat)

```

### Old data (2017-2019)    
Read from excel files in folders in each collectors' 'xlsbase'   
- Some trouble with excel file for VT42 2017 - the file is big (5 MB) but has only 1 line of data    
```{r}


read_excel_files <- FALSE
# read_excel_files <- TRUE


if (read_excel_files){
  
  #
  # 1. VR52
  #
  # dir(paste0(datafolder1, "/TilAquamonitor"))
  fn <- paste0(datafolder1, "/TilAquamonitor/VR52_CTD_2017_Til_AqM.xlsm")
  excel_sheets(fn)
  dat1a <- read_excel_droplines(fn, "Rådata",
                                first_data_row = 3, cut_columns = TRUE)

  fn <- paste0(datafolder1, "/TilAquamonitor/VR52_CTD_2018_Til_AqM.xlsm")
  excel_sheets(fn)
  dat1b <- read_excel_droplines(fn, "Rådata",
                                first_data_row = 3, cut_columns = TRUE)

  fn <- paste0(datafolder1, "/TilAquamonitor/VR52_CTD+siktdyp_2019_Til_AqM.xlsm")
  dat1c <- read_excel_til_AqM(fn, "CTD") %>%
    rename(ProjectName = PROJECT_NAME,
           StationCode = STATION_CODE,
           Date = DATE,
           Depth1 = DEPTH1,
           Depth2 = DEPTH2) %>%
    select(-DATASOURCE_NAME, -INSTRUMENT_REF, -REMARK)
  
  fn <- paste0(datafolder1, "/TilAquamonitor/VR52_CTD+siktdyp_2020_Til_AqM.xlsm")
  dat1d <- read_excel_til_AqM(fn, "CTD") %>%
    rename(ProjectName = PROJECT_NAME,
           StationCode = STATION_CODE,
           Date = DATE,
           Depth1 = DEPTH1,
           Depth2 = DEPTH2) %>%
    select(-DATASOURCE_NAME, -INSTRUMENT_REF, -REMARK)
  
  #
  # 2. VT42
  #
  # dir(paste0(datafolder1, "/TilAquamonitor"))
  fn <- paste0(datafolder2, "/TilAquamonitor/VT42_CTD_2017_Til_AqM.xlsm")
  excel_sheets(fn)
  dat2a <- read_excel_droplines(fn, "Data",
                                first_data_row = 3, cut_columns = TRUE)

  fn <- paste0(datafolder2, "/TilAquamonitor/VT42_CTD_2018_Til_AqM.xlsm")
  # excel_sheets(fn)
  dat2b <- read_excel_droplines(fn, "Rådata",
                                first_data_row = 3, cut_columns = TRUE)

  fn <- paste0(datafolder2, "/TilAquamonitor/VT42_CTD+siktdyp_2019_Til_AqM.xlsm")
  dat2c <- read_excel_til_AqM(fn, "CTD", header2_row = 3) %>%    # note header2_row
    rename(ProjectName = PROJECT_NAME,
           StationCode = STATION_CODE,
           Date = DATE,
           Depth1 = DEPTH1,
           Depth2 = DEPTH2) %>%
    select(-DATASOURCE_NAME, -INSTRUMENT_REF, -REMARK)
  
  fn <- paste0(datafolder2, "/TilAquamonitor/VT42_CTD+siktdyp_2020_Til_AqM.xlsm")
  dat2d <- read_excel_til_AqM(fn, "CTD") %>%
    rename(ProjectName = PROJECT_NAME,
           StationCode = STATION_CODE,
           Date = DATE,
           Depth1 = DEPTH1,
           Depth2 = DEPTH2) %>%
    select(-DATASOURCE_NAME, -INSTRUMENT_REF, -REMARK)

  #he
  # 3. VR31
  #
  # dir(paste0(datafolder1, "/TilAquamonitor"))
  fn <- paste0(datafolder3, "/TilAquamonitor/2017/VR31_CTD_2017_Til_AqM.xlsm")
  excel_sheets(fn)
  dat3a <- read_excel_droplines(fn, "Rådata",
                                first_data_row = 3, cut_columns = TRUE)

  fn <- paste0(datafolder3, "/TilAquamonitor/2018/VR31_CTD_2018_Til_AqM.xlsm")
  # excel_sheets(fn)
  dat3b <- read_excel_droplines(fn, "Data",
                                first_data_row = 3, cut_columns = TRUE) %>%
    # need to add StationCode
           left_join(dat_stations %>% select(StationId, StationCode), by = "StationId") %>%
    select(-StationId)

  fn <- paste0(datafolder3, "/TilAquamonitor/2019/VR31_CTD+siktdyp_2019_Til_AqM.xlsm")
  dat3c <- read_excel_til_AqM(fn, "CTD", header2_row = 3) %>%    # note header2_row
    rename(ProjectName = PROJECT_NAME,
           StationCode = STATION_CODE,
           Date = DATE,
           Depth1 = DEPTH1,
           Depth2 = DEPTH2) %>%
    select(-DATASOURCE_NAME, -INSTRUMENT_REF, -REMARK)

  fn <- paste0(datafolder3, "/TilAquamonitor/2020/VR31_CTD+siktdyp_2020.xlsm")
  dat3d <- read_excel_til_AqM(fn, "CTD") %>%
    rename(ProjectName = PROJECT_NAME,
           StationCode = STATION_CODE,
           Date = DATE,
           Depth1 = DEPTH1,
           Depth2 = DEPTH2) %>%
    select(-DATASOURCE_NAME, -INSTRUMENT_REF, -REMARK)
  
  
  # Combine data - WITHOUT THE d's: they are from 2020 anyway, so they overlap with 'dat'
  dat_old <- bind_rows(
    dat1a, dat1b, dat1c,
    dat2a, dat2b, dat2c,
    dat3a, dat3b, dat3c)

    # xtabs(~Date + StationCode, dat_old)

  # Fix the first column names
  # NOT NEEDED
  
  dat_old <- dat_old %>%
    left_join(dat_stations %>% select(StationCode, Lat, Lon), by = "StationCode")
  
  # Set -9999 to NA
  for (var in c("Saltholdighet", "Temperatur", "Oksygen", "Oksygenmetning")){
    sel <- dat_old[[var]] == -9999
    dat_old[[var]][sel] <- NA
    }
  
  saveRDS(dat_old, "Data/14_QA_CTD_2020_NorskehavSor2_dat_2017-19.rds")
  
} else {
  
  dat_old <- readRDS("Data/14_QA_CTD_2020_NorskehavSor2_dat_2017-19.rds")
  
}

if (FALSE){
  # check dates
  
  cat("-------------------------------------------------------------------------------------- \n")
  cat("dat_old: \n")
  cat("-------------------------------------------------------------------------------------- \n")
  xtabs(~addNA(Date), dat_old)
  xtabs(~addNA(Date) + StationCode, dat_old)
  
  cat("-------------------------------------------------------------------------------------- \n")
  cat("dat: \n")
  cat("-------------------------------------------------------------------------------------- \n")
  xtabs(~addNA(Date), dat)
  
}

```


## 2. Data       

### Check sample dates   
- `dat_old`= 2017-2019   
- `dat` = last part of 2019 + 2020  
```{r}
tab <- xtabs(~Date, dat_old)
missing <- is.na(dat_old$Date)
cat("-------------------------------------------------------------------------------------- \n")
cat("dat_old \n")
cat("-------------------------------------------------------------------------------------- \n")
cat("Number of dates: ", length(tab), "\n")
cat("First and last date: ", head(names(tab),1), "-", tail(names(tab),1), "\n")
cat("Missing dates: ", sum(missing), "\n")

  
tab <- xtabs(~Date, dat)
missing <- is.na(dat$Date)
cat("-------------------------------------------------------------------------------------- \n")
cat("dat: \n")
cat("-------------------------------------------------------------------------------------- \n")
cat("Number of dates: ", length(tab), "\n")
cat("First and last date: ", head(names(tab),1), "-", tail(names(tab),1), "\n")
cat("Missing dates: ", sum(missing), "\n")
  
```


### Dates and max depth of new data    
```{r, results='asis'}

dat %>%
  group_by(StationCode, Date) %>%
  summarize(n = n(), Max_depth = max((Depth1 + Depth2)/2, na.rm = TRUE),
            .groups = "drop") %>%
  arrange(StationCode, Date) %>%
  kbl() %>%
  kable_styling(bootstrap_options = "striped", full_width = F, position = "left")

# dat %>%
#   count(SampleDate)
# 
# dat %>%
#   count(StationCode, Date)

stations <- dat %>%
  count(StationCode) %>%
  pull(StationCode  )

```


## 3. Profiles of 2020 data  

### Salinity  
_NOTE: also see plots for top 20 m further down._  
   
- Gl_5: Profile looks strange at 20-30 m in April 2019 (and no pattern in temperature)  
   
- Very low salinity at 0-1 meter:     
    + Dec 2018 at Gl_3 and Gl_5: Dubious? no indication of top freshwater layer in temperature  
    + 4. April 2019 Gl_5: Dubious? no indication of top freshwater layer in temperature  
    + 21 June 2019 at Gl_6: Dubious? little indication of top freshwater layer in temperature  
    + July 2019 at Gl_2 and Gl_3: probably OK - temperature also indicates top freshwater layer  
    + 15 Aug 2019 at Gl_2 and Gl_4: Dubious? little indication of top freshwater layer in temperature  

```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

# Test one station:
# plot_ctdprofile_station("VT16", data = dat, variable = "Saltholdighet", points = FALSE)

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Saltholdighet", points = FALSE)


# stations %>% walk(plot_ctdprofile_station, data = dat, variable = "Saltholdighet", points = FALSE)

```

### Salinity top 50 m 
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

dat50m <- dat %>% filter(Depth1 <= 50)

stations %>% walk(
  plot_ctdprofile_station,
  data = dat50m, variable = "Saltholdighet", points = FALSE)


```


### Temperature    
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Temperatur", points = FALSE)


```

### Temperature top 50 m 
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat50m, variable = "Temperatur", points = FALSE)

```

### Oxygen volume      
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Oksygen", points = FALSE)

```

### Oxygen saturation        
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Oksygenmetning", points = FALSE)


```




## 4. Times series (since 2017)  

### Combined data
```{r}

dat2 <- bind_rows(
  dat %>% select(-ProjectId),
  dat_old)

# xtabs(~addNA(Date), dat2)

dat2 %>%
  distinct(Date, StationCode) %>%
  mutate(Year = year(Date)) %>%
  xtabs(~Year + addNA(StationCode), .)

# dat3b

```

### Station VR31 (SNO)      
```{r}

if (FALSE){
  # check depths
  dat2 %>%
    filter(StationCode %in% "VR31") %>%
    group_by(Date) %>%
    summarize(Max_depth = max(Depth1)) %>%
    arrange(Max_depth)
}

plot_timeseries_station(stationcode = "VR31", data = dat2, variable = "Saltholdighet", titletext = "")
plot_timeseries_station(stationcode = "VR31", data = dat2, variable = "Temperatur", titletext = "")
plot_timeseries_station(stationcode = "VR31", data = dat2, variable = "Oksygen", titletext = "")
plot_timeseries_station(stationcode = "VR31", data = dat2, variable = "Oksygenmetning", titletext = "")

```


### Station VR52 (Aquakompetanse)      
```{r}

if (FALSE){
  # check depths
  dat2 %>%
    filter(StationCode %in% "VR51") %>%
    group_by(Date) %>%
    summarize(Max_depth = max(Depth1)) %>%
    arrange(Max_depth)
}

plot_timeseries_station(stationcode = "VR52", data = dat2, variable = "Saltholdighet", titletext = "")
plot_timeseries_station(stationcode = "VR52", data = dat2, variable = "Temperatur", titletext = "")
plot_timeseries_station(stationcode = "VR52", data = dat2, variable = "Oksygen", titletext = "")
plot_timeseries_station(stationcode = "VR52", data = dat2, variable = "Oksygenmetning", titletext = "")

```

### Station VT42 (Magne Auren)      
```{r}

if (FALSE){
  # check depths
  dat2 %>%
    filter(StationCode %in% "VT42") %>%
    group_by(Date) %>%
    summarize(Max_depth = max(Depth1)) %>%
    arrange(Max_depth)
}

plot_timeseries_station(stationcode = "VT42", data = dat2, variable = "Saltholdighet", titletext = "")
plot_timeseries_station(stationcode = "VT42", data = dat2, variable = "Temperatur", titletext = "")
plot_timeseries_station(stationcode = "VT42", data = dat2, variable = "Oksygen", titletext = "")
plot_timeseries_station(stationcode = "VT42", data = dat2, variable = "Oksygenmetning", titletext = "")

```
