==> 14_QA_CTD_2021_NorskehavetSor.Rmd <==
---
title: "ØKOKYST Norskehavet Sør 2021 - QA CTDs"
author: "DHJ"
date: "13 1 2020"
output:
  html_document:
    keep_md: true
    toc: true
    toc_float: true

---

## QC done February 2022    
- CTDs with salinity + temp + oxygen      
- Excel files from OneDrive, sent by Caroline by mail (11.02.2022)  
- Companies RMS (Runde miljøsenter) and Aqua kompetanse
- Stations:    
    * VR31, VR52, VR61 - Aqua-kompetanse  
    * VT71 - RMS  

```{r, echo=FALSE, results='hide', message=FALSE, warning=FALSE}

## 1. Libraries

library(dplyr)
library(purrr)
library(ggplot2)
library(lubridate)
library(readxl)
library(tidyr)
library(RColorBrewer)
library(knitr)
library(kableExtra)

# library(niRvana)

source("12_QA_2019_from_excel_functions.R")
source("13_QA_CTD_2019_functions.R")

read_excel_add_filename <- function(filename, folder){
  fn <- paste0(folder, "/", filename)
  read_excel(fn, na = "-999") %>%
    mutate(Filename = filename)
}

# RColorBrewer::display.brewer.all()

knitr::opts_chunk$set(echo = FALSE, results = 'hold')

```


## 1. Read files   

### New data (2021)  
- sync'ed OneDrive folder from Trond Kristiansen  
```{r, echo = FALSE, results = 'hide', message = FALSE}


read_excel_files <- FALSE
# read_excel_files <- TRUE

if (read_excel_files){
  
  # Excel files from OneDrive (11.02.2022)
  # Sent by mail by Caroline and saved on PC

  datafolder <- "../OneDrive_1_11-02-2022/Aqua_kompetanse/"
  fns <- dir(datafolder, pattern = ".xlsx")
  fns <- fns[!grepl("~", fns)]
  dat1 <- fns %>% map_dfr(read_excel_add_filename, folder = datafolder)

  datafolder <- "../OneDrive_1_11-02-2022/RMS/"
  fns <- dir(datafolder, pattern = ".xlsx")
  fns <- fns[!grepl("~", fns)]
  dat2 <- fns %>% map_dfr(read_excel_add_filename, folder = datafolder)

  dat <- bind_rows(dat1, dat2)
  rm(dat1, dat2)

  # Used last year, but no StationCode in excel anymore
  # - and Date is OK already
  # dat <- dat %>%
  #     mutate(Date = as.Date(Date, origin = "1899-12-30")) %>%
  #     mutate(StationCode2 = paste0(sub("_CTD.xlsx", "", Filename, fixed = TRUE), "_", StationCode),
  #            StationCode = sub("_CTD.xlsx", "", Filename, fixed = TRUE)
  #     )
  
  dat <- dat %>%
    mutate(Date = as.Date(dmy_hms(Date)),
           StationCode = sub("_CTD.xlsx", "", Filename, fixed = TRUE),
           Month = month(Date),
           Year = year(Date))
  
  saveRDS(dat, "Data/14_QA_CTD_2021_NorskehavSor_dat.rds")
  
} else {
  
  dat <- readRDS("Data/14_QA_CTD_2021_NorskehavSor_dat.rds")
  
}

xtabs(~addNA(StationCode), dat)
xtabs(~addNA(Date), dat)

```


### Older data (2017-2020)  

```{r, message = FALSE}

dat_old1_1 <- readRDS("Data/14_QA_CTD_2020_Norskehavsor_dat_2017-19.rds") # "VR51", "VT71"    (note "wrong" filename)
dat_old1_2 <- readRDS("Data/14_QA_CTD_2020_NorskehavSor2_dat_2017-19.rds") # "VR52", "VT42", "VR31" (contains 'Fluorescense')  
dat_old2_1 <- readRDS("Data/14_QA_CTD_2020_NorskehavSor1_dat.rds")  # "VR51", "VT71"
dat_old2_2 <- readRDS("Data/14_QA_CTD_2020_NorskehavSor2_dat.rds")  # "VR52", "VT42", "VR31"

list(dat_old1_1, dat_old1_2, dat_old2_1, dat_old2_2) %>% map(names)

dat_old <- bind_rows(dat_old1_1, dat_old1_2, dat_old2_1, dat_old2_2) %>%
    mutate(Month = month(Date),
           Year = year(Date))

```


## 2. Data       

### Check sample dates   
- `dat_old`= 2017-2020   
- `dat` = last part of 2020 + 2021  
```{r}

# check dates

cat("-------------------------------------------------------------------------------------- \n")
cat("dat_old: \n")
cat("-------------------------------------------------------------------------------------- \n")
# xtabs(~addNA(Date), dat_old)
xtabs(~Year + Month + StationCode, dat_old)

cat("-------------------------------------------------------------------------------------- \n")
cat("dat: \n")
cat("-------------------------------------------------------------------------------------- \n")
# xtabs(~addNA(Date), dat)
xtabs(~addNA(Year) + StationCode, dat)
xtabs(~addNA(Month) + StationCode, dat %>% filter(Year == 2021))
# xtabs(~addNA(Date) + StationCode, dat)

cat("-------------------------------------------------------------------------------------- \n")
cat("Number of dates: ", length(tab), "\n")
cat("First and last date: ", head(names(tab),1), "-", tail(names(tab),1), "\n")
cat("Missing dates: ", sum(missing), "\n")
  
```


### Change file names for simplicity
```{r}

dat <- dat %>%
  rename(Depth1 = DEPTH1,
         Depth2 = DEPTH2,
         Saltholdighet = `Saltholdighet PSU`,
         Temperatur = `Temperatur C`,
         Oksygen = `Oksygen ml O2/L`,
         Oksygenmetning = `Oksygenmetning %`)

```


### Dates and max depth of new data    
```{r, results='asis'}

dat %>%
  group_by(StationCode, Date) %>%
  summarize(n = n(), Max_depth = max((Depth1 + Depth2)/2, na.rm = TRUE),
            .groups = "drop") %>%
  arrange(StationCode, Date) %>%
  kbl() %>%
  kable_styling(bootstrap_options = "striped", full_width = F, position = "left")

# dat %>%
#   count(SampleDate)
# 
# dat %>%2
#   count(StationCode, Date)

stations <- dat %>%
  count(StationCode) %>%
  pull(StationCode  )

```

### EXTRA: Fixing oxygen values for VT71  

- dividing values October-December by 1000 (i.e. they were given in ul [microliters] instead of ml O2/L)        

```{r}

dat_original <- dat

dat <- dat %>%
  mutate(Oksygen = case_when(
    StationCode == "VT71" & Month >= 10 ~ Oksygen/1000,
    TRUE ~ Oksygen)
  )

```


## 3. Profiles of 2021 data    

_NOTE: all plots use the data set where VR49 of 2021-10-19 has been changed to VT70, see above_   
* Exception is 'Original data' of oxygen volume   

### Salinity  
_NOTE: also see plots for top 20 m further down._  
   
- VT74:  
    - March data is duplicated (occurs twice)  
- VT70:  
    - Dec. 2020: salinity a bit rugged around 180-200 m   

```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

# Test one station:
# plot_ctdprofile_station("VT16", data = dat, variable = "Saltholdighet", points = FALSE)

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Saltholdighet", points = FALSE)


# stations %>% walk(plot_ctdprofile_station, data = dat, variable = "Saltholdighet", points = FALSE)

```

### Salinity top 50 m 
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

dat50m <- dat %>% filter(Depth1 <= 50)

stations %>% walk(
  plot_ctdprofile_station,
  data = dat50m, variable = "Saltholdighet", points = FALSE)


```


### Temperature    
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Temperatur", points = FALSE)


```

### Temperature top 50 m 
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat50m, variable = "Temperatur", points = FALSE)

```

### Oxygen volume      

- VT71 has no data zeros 

```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Oksygen", points = FALSE)

```

#### Show error in original data for VT71  

- 1000 times too high  
- fixed in section 2 'EXTRA'  

```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

plot_ctdprofile_station("VT71",
  data = dat_original, variable = "Oksygen", points = FALSE)

```

### Oxygen saturation        
```{r, echo = FALSE, fig.width=9, fig.height=7, warning=FALSE}

stations %>% walk(
  plot_ctdprofile_station,
  data = dat, variable = "Oksygenmetning", points = FALSE)


```

## 4. Times series (since 2017)  

### Combine data
```{r}

dat2 <- bind_rows(dat_old, dat) %>%
  mutate(Month = month(Date),
         Year = year(Date))

# xtabs(~addNA(Date), dat2)

dat2 %>%
  distinct(Year, Date, StationCode) %>%
  xtabs(~Year + addNA(StationCode), .)

# dat3b

```


### Station VR31       
```{r}

station <- "VR31"
plot_timeseries_station(stationcode = station, data = dat2, variable = "Saltholdighet", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Temperatur", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygen", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygenmetning", titletext = "")

```

### Station VR52       
```{r}

station <- "VR52"
plot_timeseries_station(stationcode = station, data = dat2, variable = "Saltholdighet", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Temperatur", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygen", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygenmetning", titletext = "")

```
### Station VR61       
```{r}

station <- "VR61"
plot_timeseries_station(stationcode = station, data = dat2, variable = "Saltholdighet", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Temperatur", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygen", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygenmetning", titletext = "")

```
### Station VT71       
```{r}

station <- "VT71"
plot_timeseries_station(stationcode = station, data = dat2, variable = "Saltholdighet", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Temperatur", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygen", titletext = "")
plot_timeseries_station(stationcode = station, data = dat2, variable = "Oksygenmetning", titletext = "")

```
