==> 15_QA_CTD_NordsjoNord.Rmd <==
---
title: "15_QA_CTD_NordsjoNord"
author: "DHJ"
date: "27 1 2020"
output: 
  html_document:
    keep_md: true
    toc: true
    toc_float: true
---


# QA for CTDs - ØKOKYST Nordsjøen NOrd 2019 report (January 2020)    
- Check Excel files from Anna Birgitta and Trond  
- Only CTDs
- See mail from Anna Birgitta Ledang <AnnaBirgitta.Ledang@niva.no> Sent: torsdag 23. januar 2020 14:59

   
**Veileder:**  
- `Documents\artikler\veiledere etc\02_2013_klassifiserings-veileder_.pdf`  

  
```{r, echo=FALSE, results='hide'}

## 1. Libraries

library(dplyr)
library(purrr)
library(ggplot2)
library(lubridate)
library(readxl)
library(tidyr)
library(RColorBrewer)

# library(niRvana)

source("13_QA_CTD_2019_functions.R")

# RColorBrewer::display.brewer.all()
```


## 1. Folders and file information 

```{r, results='hide'}

### Path to mother folder (must end with a slash)
basefolder <- "Datasett/Nordsjø Nord/"


# Norskehavet Sør I
folder1 <- "Hardangerfjorden"
folder2 <- "MON"
folder3 <- "Sognefjorden"


```


### Check file names, sheet names and variable names   
Makes 'fileinfo' which contains all of this info (one line per file)  
```{r, echo=FALSE}
dir(basefolder) 

folder <- folder1
# 1. Find sheet names here: (can be commented out afterwards)
# sheets_in_folder(folder, basefolder = basefolder)                                 
# 2. Enter the sheet names here:  
f1 <- vars_in_folder(folder, c(rep("Rådata", 6)), basefolder = basefolder)       

folder <- folder2
# sheets_in_folder(folder, basefolder = basefolder)                                 
f2 <- vars_in_folder(folder, c(rep("Rådata", 12)), basefolder = basefolder)       

folder <- folder3
# sheets_in_folder(folder, basefolder = basefolder)                                 
f3 <- vars_in_folder(folder, c(rep("Rådata",2)), basefolder = basefolder)       

fileinfo <- bind_rows(f1, f2, f3)

fileinfo

```

```{r check_Variables, echo=FALSE, results='hide'}

### Check which variables are in each file  

# Variables of first file:
# i <- 1
# fn <- with(fileinfo[1,], paste0(basefolder, Folder, "/", File))
# read_excel(fn, sheet = fileinfo[1,"Sheet"], n_max = 0) %>% names() %>% dput()

#
# Number of files lacking variable X  
c("StationId", "StationCode", "Date", "Depth1", 
  "Saltholdighet", "Temperatur", "Oksygen", "Oksygenmetning", "Metode") %>% map_df(files_lacking_var)
#
# All files contains StationCode


```

```{r get_all_data, echo=FALSE, results='hide'}

i <- 1
# debugonce(read_excel_droplines)
test <- read_data_fileno(i, df_fileinfo = fileinfo, first_data_row = 2)


#
# Reads all data
# This is the data used by all plots
#

datalist <- seq_len(nrow(fileinfo)) %>% 
  map(read_data_fileno, df_fileinfo = fileinfo, first_data_row = 2)

# Convert dates
for (i in seq_along(datalist))
  datalist[[i]]$Date <- as.POSIXct(datalist[[i]]$Date*24*3600, origin = "1899-12-31", tz = "GMT")

# get_data_filestation(4)

```

```{r, echo=FALSE, results='hide'}

### Get stations  
# Make 'fileinfo_stations'   
# - has one row per file and station  
# - deleting one station from one file (just 1 observation)

# Test
# i <- 1
# fn <- with(fileinfo[i,], paste0(basefolder, Folder, "/", File))
# dat <- read_excel_droplines(fn, sheet = fileinfo$Sheet[i])

#
# Makes a data frame for a given file, with one row per station
#

# Test
# get_stations(1)
# get_stations(4)

fileinfo_stations <- seq_len(nrow(fileinfo)) %>% map_df(get_stations)

# Deletes staton with 1 observation
fileinfo_stations <- 
  fileinfo_stations %>% filter(n >= 5)

fileinfo_stations

```

### Delete som negative oxygen measurements    
After plotting below  

```{r}
# Check
# datalist[[7]] %>% filter(year(Date) == 2014 & month(Date) == 6) %>% View()

# June 2014: NORD1 (7) + NORD2 (8) + OFOT2 (10) + OKS2 (12) + SG2 (14)
for (i in c(7,8,10,12,14)){
  sel <- with(datalist[[i]], year(Date) == 2014 & month(Date) == 6 & !is.na(Oksygen) & Oksygen < 0)
  cat(sum(sel), "\n")
  datalist[[i]]$Oksygen[sel] <- NA
  datalist[[i]]$Oksygenmetning[sel] <- NA
}

# June 2019: SJON2 (16, hele vannsøylen) + VT79 (20)
for (i in c(16,20)){
  sel <- with(datalist[[i]], year(Date) == 2019 & month(Date) == 6 & !is.na(Oksygen) & Oksygen < 0)
  cat(sum(sel), "\n")
  datalist[[i]]$Oksygen[sel] <- NA
  datalist[[i]]$Oksygenmetning[sel] <- NA
}

# April 2019 VT16 (19) og VT79 (20)
for (i in c(19,20)){
  sel <- with(datalist[[i]], year(Date) == 2019 & month(Date) == 4 & !is.na(Oksygen) & Oksygen < 0)
  cat(sum(sel), "\n")
  datalist[[i]]$Oksygen[sel] <- NA
  datalist[[i]]$Oksygenmetning[sel] <- NA
}

# Mai 2019 VT16 (19) og VT79 (20)
for (i in c(10)){
  sel <- with(datalist[[i]], year(Date) == 2019 & month(Date) == 5 & !is.na(Oksygen) & Oksygen < 0)
  cat(sum(sel), "\n")
  datalist[[i]]$Oksygen[sel] <- NA
  datalist[[i]]$Oksygenmetning[sel] <- NA
}

# Feb 2017: OKS2 (12)
for (i in c(12)){
  sel <- with(datalist[[i]], !is.na(Oksygen) & Oksygen < 0)
  cat(sum(sel), "\n")
  datalist[[i]]$Oksygen[sel] <- NA
  datalist[[i]]$Oksygenmetning[sel] <- NA
}

```

## 2. Plot casts

Plot all casts  
- Also check max depth    
**Notes**  
- NB: Samme StationCode på mange av filene (eks. StationCode = 69165 for alle Hardangerfjordfilene VT52, VT53, etc.). Se over.     
- VT69 har stor variasjon i dybde: 12 meter i feb 2017, rundt 40-50 meter mars 2017 til okt. 2018, så 70-90 meter (opptil 95 meter) vinteren 2018-2019 før det går tilbake til ca 50 meter (obs: kun 32 meter i aug 2018): Hva har skjedd her?  
- 35-50 meter variasjon i maks.dybde for noen av de andre Hardangerfjordfilene, men det er vel greit siden disse er på stort dyp  
- MON/NORD1: lav maksdybde i mars 2013 (229 m) og juni 2014 (222 m), skal være 285-290 m  
- MON/NORD2: Ligger oftest på 225-230 m, men har ekstra stor maksdybde i mars 2013 (289 m), så denne må være et annet sted  
- MON/OFOT1: Typisk 230-240 meter, men store avvik begge veier: små dybder i sept 2013 (156 m), juli 2018 (75 m), okt 2018 (36 m), og veldig store dybder ( 433-436 meter) i feb 2014, april 2018 og jan 2019.   
- MON/OFOT2: Normalt 430-440 m. Lave dybder (<300 m) i des 2013, feb 2014, juni 2014 aug 2016, april 2018, jan 2019  
- MON/OKS1: Ganske stort avvik i des. 2013 (122 m istedet for normalt 150-190 m)   
- MON/OKS2: Ganske stort avvik i des. 2013 (135 m istedet for normalt 175-195 m)  
- MON/SAG1: Normalt rundt 600 m, store avvik i des 2013 (436 m) og mars 2018 (343 m)   
- MON/SAG2: Normalt rundt 340-350 m, store avvik i des 2015 (253 m)   
- Sognefjord/SJON1: OK   
- Sognefjord/SJON2: april 2017: 371 m mot normalt ca 620   
- Sognefjord/TYS1: nov-des 2013: <350 m mot normalt 720 m, også 571 m i sept 029
- Sognefjord/TYS2: normal dybde ca 600 m, lav (<450 m) nov-des 2013, høy (708 m) i apr 2013  
- Sognefjord/VT16: >1250 m normalt, store avvik i jan 2019 (893 m) og aug 2019 (496 m)   

```{r, fig.width = 9, fig.height = 7}

plot_cast_all()

```

## 3. Plot parameters

### Plot salinity time series

```{r}
plot_timeseries_all("Saltholdighet")
```

### Plot salinity profiles

```{r, fig.width=9, fig.height=7}
plot_ctdprofile_all("Saltholdighet")
```

### Plot temperature time series
```{r}
plot_timeseries_all("Temperatur")
```

### Plot temperature profiles   
```{r, fig.width=9, fig.height=7}
plot_ctdprofile_all("Temperatur")
```

### Plot oxygen conc time series    

Veilederen sier:   

> Oksygenmålingene og klassifisering er knyttet til maksimalt dyp,   
> men det anbefales at målingene foretas som vertikale profiler   
> for å avklare om større deler av vannsøylen har reduserte oksygenmengder.  
   
- VT52: Oksygenkonsentrasjon er ofte lavere rundt 100-200 m dyp enn ved bunnen (og lavere enn 4.5 i noen måneder)   
- VT52: Oksygenkonsentrasjon er ofte veldig nær grensa på 4.5 i en stor del av vannsøylen    
- OFOT1: Oksygenkonsentrasjon er ofte lavere rundt 100 m dyp enn ved bunnen
- "Hakkete" oksygenprofiler i des. 2018 (VT53, VT74) og mai 2019 (alle Sognefjordsdata). Ser tvilsomt ut, kan være feil på sensor.  
-  Vær oppmerksom på tidstrender i noen av datasettene og at bra tilstand i 2017-2018 kan "maskere" dårlig tilstand i de siste målingene (2019) og eldre målinger (2013-2014). Bør bemerkes.    

```{r}
plot_timeseries_all("Oksygen")

```

### Plot oxygen conc. profiles     
"Hakkete" oksygenprofiler i des. 2018 (VT53, VT74) og mai 2019 (alle Sognefjordsdata)
```{r, fig.width=9, fig.height=7}
plot_ctdprofile_all("Oksygen", limits = 4.5)
```



### Plot oxygen saturation time series

```{r}
plot_timeseries_all("Oksygenmetning")
```

### Plot oxygen saturation profiles

```{r, fig.width=9, fig.height=7}
plot_ctdprofile_all("Oksygenmetning", limits = c(50,65))
```
