==> 17_QC_Nutrients_2021_Norskehavet_Sor.Rmd <==
---
title: "QC 2021 Norskehavet Sør, nutrients + Chl"
author: "DHJ"
date: "10.02.2022"
output: 
  html_document:
    toc: true
    toc_float: true
    keep_md: true
---

## QC of Norskehavet Sør  

- RMS (Runde) stations:  
    * VT71 Skinbrokleia  
    * VR51 Krossen  
- Aquakomp stations:
    * VR31 Tilremfjorden   
    * VR52 Broemsneset
    * VR61 Frohavet Sør   
    
- **Based on downloaded data from Aquamonitor**  

**Veileder:**  
- `Documents\artikler\veiledere etc\02_2013_klassifiserings-veileder_.pdf`  
- Chlorophyll: page 91.   
"I SørNorge (til Stadt) anbefales det at innsamlingen starter i februar og avsluttes ved utgangen av oktober."   
"Nord for Stadt anbefales det at innsamlingsperioden strekker seg fra mars til og med september."   
90% percentiles should be used   
  
- Also see calculation example on on page 187-189, and the comment on page 189:  
"Det gjøres oppmerksom på at dataprogrammene som beregner 90-persentil, gjerne benytter ulike metoder". We use the Excel method PERCENTILE.EXC, corresponding to type 6 quantiles in R  
  
- Nutrients: page 102 in veileder  
Vinter er november til og med februar   
Sommer er juni til og med august  

  
```{r, echo = FALSE, results='hide', warning=FALSE, message=FALSE}

### 0. Libraries   

library(dplyr)
library(purrr)
library(ggplot2)
library(lubridate)
library(readxl)
library(tidyr)
library(knitr)         
library(RColorBrewer)

# library(niRvana)

source('11_Read_from_Nivabasen_functions.R')  # Used to read from Nivadatabase

source("12_QA_2019_from_excel_functions.R")
source("13_QA_CTD_2019_functions.R")          # plot_ctdprofile_station

library(niRvana)

knitr::opts_chunk$set(results = 'hold', echo = FALSE)

# RColorBrewer::display.brewer.all()
```
### Projects  
```{r}

df_proj_all <- niRvana::get_projects()

df_proj <- df_proj_all %>%
  filter(grepl("økokyst", PROJECT_NAME, ignore.case = TRUE),
         grepl("2021-2025", PROJECT_NAME, fixed = TRUE))

df_proj

```

### Stations  
```{r}

proj_id <- df_proj %>%
  filter(PROJECT_NAME == "ØKOKYST Norskehavet Sør 2021-2025") %>%
  pull(PROJECT_ID)

df_stations_all <- get_stations_from_projectid(proj_id)

df_stations <- df_stations_all %>%
  group_by(STATION_ID) %>%
  summarise(
    across(c(PROJECT_ID, STATION_CODE, STATION_NAME), 
           .fn = ~ paste(unique(.x), collapse = ","))
  ) %>%
  arrange(STATION_CODE)


```


### Get chemistry data up to 2020    

```{r}

df_chem_2020 <- readRDS("Data/17_2020_NorskehavetS2.rds")   

df_chem_2020 <- df_chem_2020 %>%
  mutate(Date = floor_date(SAMPLE_DATE, unit = "day"))

cat("Observations per parameter \n")
xtabs(~STATION_CODE + PARAM_NAME, df_chem_2020)  

cat("Number of dates per year \n")  
df_chem_2020 %>%
  distinct(STATION_CODE, Year, Date) %>%
  xtabs(~Year + STATION_CODE, .)  

```


### Get chemistry data for 2021      

```{r}

# Read from NIvabase (1 minute) or read from saved data?
read_from_nivabase <- FALSE
# read_from_nivabase <- TRUE


if (read_from_nivabase){
  
  # Get all 2021 data
  
  stations <- df_stations %>% pull(STATION_CODE)
  
  # debugonce(get_water_chemistry)
  # debugonce(get_water_samples)
  
  df_chem_2021 <- get_water_chemistry(
    stationcodes = stations,
    years = 2021) %>%
    mutate(Year = year(SAMPLE_DATE),
           Month = month(SAMPLE_DATE),
           Date = floor_date(SAMPLE_DATE, unit = "day"))
  
  # Uncomment to save (overwrite) data:
  saveRDS(df_chem_2021, "Data/17_2021_NorskehavetS.rds")  
  
} else {

  df_chem_2021 <- readRDS("Data/17_2021_NorskehavetS.rds")  
  
}

stations <- unique(df_chem_2021$STATION_CODE)

cat(nrow(df_chem_2021), "rows of data downloaded \n\n")  
cat("Stations:", paste(unique(df_chem_2021$STATION_CODE), collapse = ", "), "\n\n")
cat("Parameters:", paste(unique(df_chem_2021$PARAM_NAME), collapse = ", "))

```

## Overview tables  

#### Methods  
```{r}

df_chem_2021 %>%
  count(PARAM_NAME, PARAM_UNIT, METHOD_ID, CONVERSION_FACTOR, NAME, UNIT)

```

#### Parameters per station  
```{r}

cat("Observations per parameter \n")
xtabs(~STATION_CODE + PARAM_NAME, df_chem_2021)

cat("\n")
cat("Dates per parameter \n")
df_chem_2021 %>%
  distinct(STATION_CODE, PARAM_NAME, Date) %>%
  xtabs(~STATION_CODE + PARAM_NAME, .)

```

#### TOTP vs TOTP_P in VT71   
* TOTP_P is much smaller than TOTP  
```{r}

df_chem_2021 %>%
  filter(STATION_CODE == "VT71",
         grepl("TOTP", PARAM_NAME)) %>%
  xtabs(~Date + PARAM_NAME, .)

# Example (June)
check <- df_chem_2021 %>%
  filter(STATION_CODE == "VT71",
         grepl("TOTP", PARAM_NAME),
         Month == 6) %>%
  select(DEPTH1, PARAM_NAME, VALUE) %>%
  pivot_wider(names_from = PARAM_NAME, values_from = VALUE)

check
check[,3]/check[,2]

```

#### Depths per station/parameter     
```{r}

df <- df_chem_2021 %>%
  distinct(STATION_CODE, PARAM_NAME, DEPTH1, Date)

tab <- xtabs(~STATION_CODE + PARAM_NAME + Date, df)

cat("==== Max no depths ==== \n")
apply(tab, 1:2, max)

cat("\n\n")
cat("==== Min no depths ==== \n")
old.o <- options(warn = -1)   # turn off warnings
apply(tab, 1:2, function(x) min(x[x > 0]))
options(old.o)                # turn warnings back on

```

#### Dates  
```{r}

# df_chem_2021 %>%
#   distinct(STATION_CODE, Date) %>%
#   count(STATION_CODE)

cat("Number of dates per month \n")
tab <- df_chem_2021 %>%
  distinct(STATION_CODE, Month, Date) %>%
  xtabs(~Month + STATION_CODE, .)
tab

cat("\n\n")
cat("Number of dates per year \n")
apply(tab, 2, sum)

```
#### Dates for CTD vs chemistry  
```{r}

vars_ctd <- c("O2", "O2 (metning)", "Salinitet", "Temperatur", "Turbiditet")

vars_lab <- c("Siktdyp", "Nitrogen part", "POC", "TOTP_P", "DOC", "TSM", 
"TOTP", "TOTN", "NH4-N", "NO3+NO2-N", "PO4-P", "SiO2", "KlfA")

cat("Number of CTD dates per month \n")
tab <- df_chem_2021 %>%
  mutate(STATION_CODE = factor(STATION_CODE)) %>%  # in order to show all stations  
  filter(PARAM_NAME %in% vars_ctd) %>%
  distinct(STATION_CODE, Month, Date) %>%
  xtabs(~Month + STATION_CODE, .)
tab

cat("\n")
cat("Number of lab-variable dates per month \n")
tab <- df_chem_2021 %>%
  mutate(STATION_CODE = factor(STATION_CODE)) %>%  # in order to show all stations  
  filter(PARAM_NAME %in% vars_lab) %>%
  distinct(STATION_CODE, Month, Date) %>%
  xtabs(~Month + STATION_CODE, .)
tab


```

### Set df_chem  
```{r}

df_chem <- df_chem_2021

```


### Check for duplication of data    
```{r}

check <- df_chem %>%
  group_by(STATION_CODE, SAMPLE_DATE, DEPTH1, PARAM_NAME) %>%
  summarise(n = n(), .groups = 'drop') %>%
  filter(n > 1)

if (nrow(check) > 0)
  stop("Some combinations of STATION_CODE, SAMPLE_DATE, DEPTH1, PARAM_NAME occurs >1 time!")

df_chem %>%
  count(PARAM_NAME, NAME)

```

### Reformat data    
Formatting the data in 'wide' format (one column per paremeter) as shown below  
```{r}

dat <- df_chem %>%
  select(STATION_CODE, SAMPLE_DATE, DEPTH1, PARAM_NAME, VALUE) %>%
  tidyr::pivot_wider(names_from = "PARAM_NAME", values_from = "VALUE") %>%
  rename(StationCode = STATION_CODE, 
         Date = SAMPLE_DATE, 
         Depth1 = DEPTH1) %>%
  mutate(Year = year(Date),
         Year =  factor(Year),
         Month = month(Date),
         Month =  factor(Month, levels = 1:12),
         `Nitrate + ammonium` = `NO3+NO2-N` + `NH4-N`) %>%
  arrange(StationCode, Date, Depth1)


head(dat, 4)

```



## Profile plots

## Plots for each variable {.tabset}  

### Chla   
```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "KlfA"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

stations %>% walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```

### NH4-N

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "NH4-N"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

stations %>% walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```



### NO3+NO2-N

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "NO3+NO2-N"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

stations %>% walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```


### TOTN  

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "TOTN"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

stations %>% walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```

### TOTN -2020  

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "TOTN"

df_chem_2020 %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_grid(vars(Year), vars(DEPTH1), labeller = label_both) +
  labs(title = var)

```

### Nitrate+nitrite, ammonium and TotN  
```{r}

stations %>% 
  walk(
    plot_ctdprofile_station_multi, 
    data = dat,
    variables = c("NO3+NO2-N", "NH4-N", "TOTN"),
    points = TRUE, titletext2 = "Nitrate+nitrite, ammonium and TotN"
)

```


### PO4-P

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "PO4-P"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

stations %>% walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```


### TOTP

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "TOTP"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

stations %>% walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```





### SiO2

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "SiO2"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

stations %>% walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```


### POC

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "POC"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

df_chem %>% filter(PARAM_NAME == var) %>%
  pull(STATION_CODE) %>% unique() %>% 
  walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```

### DOC

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "DOC"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

df_chem %>% filter(PARAM_NAME == var) %>%
  pull(STATION_CODE) %>% unique() %>% 
  walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```


### TSM

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "TSM"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

df_chem %>% filter(PARAM_NAME == var) %>%
  ggplot(aes(Month, VALUE, color = STATION_CODE)) +
  geom_line() + geom_point() + 
  facet_wrap(vars(DEPTH1), labeller = label_both) +
  labs(title = var)

plot_statistics_variable(df_chem, var)

df_chem %>% filter(PARAM_NAME == var) %>%
  pull(STATION_CODE) %>% unique() %>% 
  walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = TRUE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```


### Salinitet

```{r, echo = FALSE, fig.width=10, fig.height=4.5, warning=FALSE}

var <- "Salinitet"

max_depth <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(DEPTH1) %>% max() %>% pretty() %>% .[2]

max_value <- df_chem %>% filter(PARAM_NAME == var) %>%
  pull(VALUE) %>% max() %>% pretty() %>% .[2]

# NOT SUITABLE HERE (just one station) (also: one plot per depth)
# df_chem %>% filter(PARAM_NAME == var) %>% View()
#   filter((DEPTH1-0.5) %in% c(0, 20, seq(50,250,50), seq(300,800,100))) %>%
#   ggplot(aes(Month, VALUE, color = STATION_CODE)) +
#   geom_line() + geom_point() +
#   facet_wrap(vars(DEPTH1), labeller = label_both) +
#   labs(title = var)

plot_statistics_variable(df_chem, var)

# NOTE: points = FALSE sincce it's CTD
df_chem %>% filter(PARAM_NAME == var) %>%
  pull(STATION_CODE) %>% unique() %>% 
  walk(plot_ctdprofile_station, data = dat, 
                  variable = var, points = FALSE, year_by_month = TRUE, 
                  maxdepth = max_depth, maxvalue = max_value)

```



## Check older Secchi depth for VR52 Broemsneset        

```{r, fig.width=10, fig.height=8}

# Read from NIvabase (1 minute) or read from saved data?
read_from_nivabase <- FALSE
# read_from_nivabase <- TRUE


if (FALSE){
  
  df_secci_vr52 <- get_water_chemistry(
    stationcodes = "VR52",
    parameters = "Siktdyp") %>%
    mutate(Year = year(SAMPLE_DATE),
           Month = month(SAMPLE_DATE),
           Date = floor_date(SAMPLE_DATE, unit = "day"))
  
  df_N_vr52 <- get_water_chemistry(
    stationcodes = "VR52",
    parameters = c("NO3+NO2-N", "NH4-N", "TOTN")) %>%
    mutate(Year = year(SAMPLE_DATE),
           Month = month(SAMPLE_DATE),
           Date = floor_date(SAMPLE_DATE, unit = "day"))
  
  ggplot(df_N_vr52, aes(DEPTH1, VALUE, color = PARAM_NAME)) +
    geom_point() +
    geom_line() +
    facet_grid(vars(Year), vars(Month)) +
    coord_flip()

}

stations <- unique(df_chem_2021$STATION_CODE)

cat(nrow(df_chem_2021), "rows of data downloaded \n\n")  
cat("Stations:", paste(unique(df_chem_2021$STATION_CODE), collapse = ", "), "\n\n")
cat("Parameters:", paste(unique(df_chem_2021$PARAM_NAME), collapse = ", "))

```
